camera-based teleop
Transform ^0_4 is now calculated from fw kin for arms with camera
tools. This is used to transform incoming teleop packets so that
master movement along the table axis is translated to purely
insertion axis (z4, z3) and only the roll along this axis
is translated to camera roll. up/down/left/right should map to
equivalent motions in camera frame, but need to be tested and
altered based on orientation of camera.