CRTK state and status interface
a new CRTK state class holds the CRTK-defined state and status flags.
The state is mapped from the RAVEN runlevel and published as a state
message (pending update from CRTK-meeting). (Pedal state seems to be
stored in 'surgeon_mode' variable instead of using runlevels 2,3.
This should be investigated further, but has not impacted function-
ality yet.)