receiving servo_cr commands
servo_cr commands are being stored in the newtowrk side and probably
the device side, too. We made the decision that things should be
stored in CRTK frames and units (kg, N, m, s) until they're passed
into the RAVEN structures.  They aren't being pushed to the robot
structur yet, though. Patience, please.

We're sending things around as pointers more so that things are
changed permanently. So that's nice.